+++ pos-server/src/main/java/com/saulpos/server/receipt/repository/ReceiptHeaderRepository.java
@@ -1,0 +1,7 @@
+package com.saulpos.server.receipt.repository;
+
+import com.saulpos.server.receipt.model.ReceiptHeaderEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReceiptHeaderRepository extends JpaRepository<ReceiptHeaderEntity, Long> {
+}